CA-227062: remove spurious backtrace
When calling `xe host-*` with the `host` param, we were first checking
if host represented a `uuid` using `host.get_by_uuid`, and only after a
failure we were probing for the (host)name (using the `name-label` field
for the search). This was leaving a spurious backtrace in the logs when
`host` was not a uuid, and could erroneusly fail in case the `name-label`
was modified.

The same issue was present when using the `sr` and `vm` selector.

The fix involves using the filtering capabilities of
`get_all_records_where` to select an appropriate subset of records
that present the selected value as `name-label` or `uuid`.

In the `host` selector case, the `hostname` field is now used along with
`uuid` and `name-label` ones, solving the potential out-of-sync issue
mentioned above.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>