CA-228680: Split general-purpose assert into 2 separate asserts
assert_power_state_is and assert_power_state_in are currently used in
different ways; sometimes as a precheck to an operation, other times as
a postcheck, and, in 2 instances, as part of another assertion (where
those are also used as prechecks).

The assert has been split into 2 separate assert functions, where one
is supposed to be used as a precheck (which throws VM_BAD_POWER_STATE
to signify that the user has attempted something silly) and another to
be used as a postcheck (which throws INTERNAL_ERROR to signify that
something has gone wrong during an operation and lead to the VM ending
up in an unexpected power state).

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>