CA-233915 turn post-op assertions that VM is "Running" into warnings
Most VM operations that change the power state of a VM include
assertions that the power state is indeed as expected after that
operation. Such a check can issue a false alarm in some cases and this
commit removes it from operations that change the power state to
"running" but still logs any unexpected power state:

    start
    unpause
    reboot
    resume

The problem was that the assertion was too naive when changing the
state into running: when a VM quickly changes its state once it is
running (e.g., by rebooting) the assertion can fail because it can be
too late and would find the VM not running as expected.

Assertions after operations that change the state not to running are
kept by this commit.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>