CA-234494: introduce default templates field
The mechanism used to generate and manage default templates was based on
a custom optional key in the `other-config` field. One of the by-products
of this is that when default templates are generated they are at first
simple templates, and inherit their allowed operations. The set of
operations for default templates, however, should not include migrations.

With this change we are introducing a new backward compatible field, called
`is-default-template`. When its value is set, the VM becomes (a template
and) a default template, and the allowed operations are updated.

This gets rid of the issue and polishes the api to manage and create
default templates making them first class members of the API. In the
future it would be good to get rid of the backward compatibility checks
and sets on the other-config field.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>