CA-234510: update db before archiving rrd
The `Rrdd.archive_rrd` operation could wait for a mutex or be stuck in IO operations for some time. In rare cases, this means that the power state (and allowed operations) informations in the database could be wrong for a long amount of time.

By inverting the order of operations, we make sure that the potentially slow operation only happens  after the database state has been completely updated.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>