CA-236444: xapi reports WLB consult error when host enters maintenance mode
xapi does not understand the response messagefor the host
recommendation from wlb. In the response message, the
CanPlaceAllVMs node has two siblings, but xapi does not know.
This commit checks now whether the response message
is in any of the CanPlaceAllVMs nodes.

CA-236444: make WLB recommendations more robust

For WLB recommendations Xapi looks for a specific node in the
XML code returned by WLB. This commit makes looking for that
node more robust such that it no longer fails. In particular,
Xapi now searches all siblings of a node rather than expecting
the desired node to be in a fixed position (which might change).

Signed-off-by: Lin Liu <lin.liu@citrix.com>