CA-241704: switch from 'yum install' to 'yum upgrade'
If an existing update is upgraded (e.g. a device driver) it increases
its version number but retains its original name. When yum is asked to
install the update package it reports it as a no-op as a package of
that name is already present.

Using 'yum upgrade' instead makes upgrades behave as expected. Initial
installs continue to work as well.

Signed-off-by: Simon Rowe <simon.rowe@eu.citrix.com>