CA-241301: Don't push RRD for VM if it is not running
The function `Rrdd_proxy.push_rrd` is called from message_forwarding at the end
of a VM start or resume. It looks up the VM's host of residence, assuming the
VM is now running. If it has already crashed (for whatever reason), then the VM
isn't resident anywhere, and we should not attempt to push the RRD. The
start/resume call will still succeed in this case, because the VM crashed
_after_ starting/resuming it.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>