CP-20432: Remove some host params that have to be localhost
In some functions that took a host parameter that had to be the
local host, I have removed that parameter and made the functions
call Helpers.get_localhost instead.

That get_localhost function has not been changed here: it looks
as if it could be optimised by using Xapi_globs.localhost_ref when
it has been set, but that approach causes failures in our current
test code.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>