Stop using PCI.functions field
The `PCI.functions` field in the DB is an internal-only field that has never
really been used. It was hardcoded to always be `1L`, and its intended use has
now been replaced by the new `PCI.physical_function` and
`PCI.virtual_functions` fields.

The field was used for GPU passthrough on the capacity-checking code paths.
This patch assumes that GPU passthrough always means passing through an entire
GPU. Virtual functions are irrelevant in this case, so the code is simplified
to just check whether the PCI device is in use by a VM or not.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>