Add SR-IOV awareness to Xapi_vgpu_type.requires_passthrough
With the addition of MxGPU VGPU types, we need to distinguish between VGPU
types that need a full PGPU passed-through to the VM (the traditional
passthrough types) and ones that require passing through of a virtual function
(SR-IOV).

This patch updates the Xapi_vgpu_type.requires_passthrough function to provides
this more fine-grained information.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>