Vgpuops: factor out GPU allocation code
The main function in Vgpuops, `create_vgpus`, is split into two cases: one case
is based on PCI assignment for GPU passthrough, the other is for vGPU. Both
cases use the same allocation helper functions, but in slightly different ways.

The main difference is that the vGPU allocation algorithm takes a parameter
that can be set to `depth_first` or `breadth_first`, which controls in which
way to "fill" up the pGPUs. This is irrelevant to the passthrough case, because
the remaining capacity is binary, but using it causes no harm, and simplifies
the code.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>