CP-20787: Call mxgpu_vf_setup once if no virtual functions are found
If we find no free virtual functions for a given MxGPU device, even though we
have completed the capacity planning phase, the most likely reason is that the
driver has not yet been loaded.

We call Xapi_pgpu.mxgpu_vf_setup once, which loads the driver and creates PCI
objects in the DB for any newly found VFs. If we still find no free VFs after
that, then we may have mis-assessed the free capacity (an internal error).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>