Use PCI.scheduled_to_be_attached_to to avoid race when allocating VFs
This field is used to mark a PCI device as reserved by a VM at the time it is
allocated and before the device has actually been attached to the VM after it
has started. It can then no longer be assigned to another starting VM.

The field is cleared once the PCI device has eventually been attached, and in
case the VM's state is reset (e.g. when the server restarts, for cloned VMs,
etc.)

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>