[CA-203227] Improve VM shutdown performance for many VDIs
The reason that VM shutdown was very slow with many VDIs is because the destroy_sr code becomes slow with a large number of VDIs attached : for every VDI it attempts to lock it, then it checks if it has the datapath being removed. In practice this is unnecessary, and on this basis we can drastically improve the performance. This code makes 2 assumptions, which it verifies - raising an exception if they become untrue.

1) A datapath is only ever on zero or one VDIs.
2) The VDIs containing a datapath aren't changed by another thread while the destroy_sr call works.

Assumption 2 means that we don't need to lock every VDI to check if it has the datapath, we can just do the check at the beginning - checking without locking is very cheap, it's the lock contention that causes poor performance. Assumption 1 allows further simplification meaning we can work with Options instead of lists. Note that the signature of destroy_sr changes due to assumption 1, it now returns an Option (None or Some exception) instead of a list of exceptions (since it only works on the Option (vdi, vdi_t) - not all of the VDIs.

Also note that the debug code adds to the runtime complexity of this function (most notably we check every VDI again to verify assumption 2), but this code is no longer the bottleneck and observed shutdown performance wouldn't improve drastically if that was removed.

*Performance numbers (shutdown VM with 255 VDIs):*
Before: ~3m30s
After: ~1m40s
For comparison, starting this VM takes ~1m50s.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>