Fix VGPU.currently_attached setting in Xapi_xenops.update_pci
This looks like a case where a couple of lines were thought to be outside the
preceding match, but were actually inside the final case. The conditionals in
place made sure that this code would never be executed:

  match vgpu_opt with
  | Some vgpu -> (* stuff *)
  | None -> ();
    Opt.iter (fun vgpu -> (* dead code, since vgpu_opt = None)) vgpu_opt

If we have any open tickets about `VGPU.currently_attached` not set for
passthrough VGPUs, then this is why.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>