MxGPU: set VGPU.{resident_on;currently_attached} in just one place
MxGPU VGPUs have both a PCI device and a VGPU "device" in the xenopsd metadata
for the VM, whereas older VGPU types have just one of the two: PCI for
passthrough types, and VGPU for emulated types.

The get_state calls in the `update_pci` and `update_vgpu` functions in
`Xapi_xenops` both drive the `VGPU.{resident_on;currently_attached}` fields. So
far, at most one of these two `update_` functions would be called per VM, but
for the new case both will be called, and both will attempt to set the DB
fields. It's really the PCI state that matters for MxGPU, so we add a
conditional to `update_vgpu` to not update the fields for passthrough VGPUs.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>