CA-245402/PAR-213: Use the right bridge in VIF.move
Commit 39a09cbd move the responsibility for adding VIFs to bridges from
vif-real to setup-vif-rules. The inputs to these scripts are a little
different. For example, vif-real takes the bridge name from xenstore, while
setup-vif-rules reads it directly from xenopsd's runtime DB (JSON files on
disk). This means that we now need to make sure that this DB is up-to-date
before making any calls to the scripts. The same pattern was already used by
VIF.set_locking_mode.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>