CA-242706,CA-193907: regard IP changed at startup
in the Xapi_mgmt_iface.on_dom0_networking_change function, even if the
database already holds the correct value, because it gets set earlier in
the startup sequence. For that reason, the function did not detect a
changed IP address at startup, and did not execute the necessary
triggers. This commit changes the situation so that these "triggers" are
always executed at startup, even if the IP address did not change
between reboots.

The fix for CA-224335 moved the update_getty trigger into a branch in
this function, to only call it when the IP is actually changed, and
reduce the number of times this function is called, because that was
essentially the problem - it was called too many times. I think this
commit will not reintroduce this problem, because for all the other
calls of Xapi_mgmt_iface.on_dom0_networking_change, this function is
still guarded by the "if", it is only during startup that it's
unconditionally executed.

This commit is just a superficial workaround that does not fix the
underlying problem - that the host's ip address and the functions and
database records depending on it are not tied together and are not
guaranteed to be executed/updated when the former changes.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>