CA-247700: Don't GC VBDs that are currently attached
This prevents us getting into the situation described in CA-247700,
that we had 2 datapaths to the VM for xvda for different VDIs. The
scenario was that a VHD file was removed from the SR somehow (out
of band), which caused the SR-scan to remove it from xapi's db,
which then caused the VBD to be reaped by the GC. A subsequent VBD
was created for the same device and then plugged. Let's not even
have that a possibility - keep the attached VBD around until the
VM is shut down, at which point it will be GC'd.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>