[CP-21359] Remove the multiplier on update size calculations
See the ticket for full explanation, but for Ely updates we will use the sum of the sizes of the packages in an update as the installation size, instead of the delta of the installed size from the last major release. Since that already represents the worst case space needed we dont't need to multiply it to calculate the required space now.

Justification for keeping the disk-space get_required_space method (which now just returns its args["size"]) is to avoid forcing XenCenter/the client to be aware of this change - they can continue calling the plugin with the new update size and it'll behave as expected.

-

More info on the new size:
The delta being used currently represents the extra space required on the host *after* the update is installed, and we previously used a multiple of that to try to approximate the space needed for the reboot (which previously needed to extract/install all in dom0, but now doesn't). However the space required for the actual install is equal to the size of the packages being installed, since we can't unmap the old version until the new one is in place and the service etc. restarted. So this is the actual worst case install size, and using it with a multiple of 1 is much more accurate than using a somewhat arbitrary multiple of the deltas.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>