Better interface
There are now three types - 'id', 'task_handle' and 'tasks'. The
'tasks' record is a set of tasks, an 'id' can be used to reference
a task within a task list, and a 'task_handle' is used for all
operations on tasks. Once you have a 'task_handle' you don't need
to reference the 'tasks' record again, the 'task_handle' knows which
set of tasks it came from.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>