CA-244657: Allow xapi to overwrite the cancellation policy at higher level
Currently, many xenopsd tasks support cancellation. The cancellation at xenopsd
level usually means just abruptly stopping the low-level operation and do some
limited cleanup.  Sometime, the result is not necessarily a user would expect
from higher level point of view. For example, some operations are destructive,
so that a cancellation won't leave the objects involved in their original
status but some status unexpected to the user. Also, some operations are
implemented as combinations of a sequence of different tasks underlying, so
cancellations at different stages will lead to different results/behaviors,
which could further confuse the users.

Currently when calling from xapi down to xenopsd, the cancellability of the
xapi task is simply bound to that of the underlying xenopsd tasks. This patch
adds an extra flag to allow overwriting this if necessary.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>