CA-244657: Disable cancellation for some xapi tasks
Currently, operations like VM suspend/checkpoint/revert are based on some
destructive low-level implementation. When they are cancelled in the middle of
the process, the status of involved VM might not be what a user would expect.
Given that these operations are not particularly time consuming, it's better to
leave them running through without interruption. And if a user is not happy
with the output, for these operations there is always a way to revert the
output afterwards.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>