.travis.yml: explicitly set XS branch name
Because the git branch name does not match the XS branch name in this
case. This will fix the Travis build as Travis will now use the correct
RPMs.

This commit depends on corresponding changes in xenserver-build-env,
commit xenserver-build-env#0f1f452, and commit
xenserver-build-env#2c4b92a, which makes it possible to use a git branch
name that is different from the XS branch, and explicitly set it in the
GIT_BRANCH environment variable.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>