CP-20681: introduce VGPU_DESTINATION_INCOMPATIBLE error
This will be raised if the target host has pGPUs in the right GPU group,
but NVidia's compatibility checker raises a problem. The error has one
argument: a string that describes the reason for the incompatibility,
which can currently be only `host-driver` or `unknown`.

An additional NVIDIA_TOOLS_ERROR is introduce to signal issues with the
Nvidia tool libraries on specific hosts (e.g. missing or outdated library).

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>