CA-251254: Force a xenopsd VM metadata update before a pool-migration
This is done after the assignment of a new PGPU for the VM (in case the VM has
a VGPU). The metadata update includes the PCI address of the PGPU on the
destination host, which gets passed by xenopsd on the sender to xenopsd on the
destination. If we don't do this, then the destination's xenopsd may have the
wrong PCI address when setting up the VGPU.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>