CP-20679: Optimise some bits, simplify others
A small optimisation: pass the appropriate value of
compatibility_metadata to the PGPU constructor, rather than creating
the object and then populating the field afterwards.

Doing this involved changing some of the underlying functions, and
we make fewer database lookups now.

Also the code now passes the correct debug-string to the xcp-idl
xmlrpc call that fetches the gpgu metadata.

Also removed some needless detail from a specified return-type in an
mli file.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>