CP-20679: cope when pgpu metadata is unobtainable
The code that fetched the pgpu metadata from the Nvidia driver was
assuming that this would always work for every GPU made by Nvidia.

In fact only some Nvidia GPUs support the feature, and even where
the feature is available, and the Nvidia driver has been installed
on the host, that Nvidia driver has not yet been loaded when xapi
starts on host boot.

Therefore this commit catches the exception
Gpumon_interface.NvmlInterfaceNotAvailable
and the exception handler returns and empty list for the metadata.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>