Add a VM.assert_can_migrate_sender internal call
VM.assert_can_migrate forwards VM.assert_can_migrate_sender to the sending host
of a proposed live migration (where the VM is currently running) to run
pre-migration checks that can only be run on that host. The remaining checks
(currently all of them) can be done using just the information in the database,
and are therefore run on the master.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>