CP-20682: Assert free vGPU capacity on receiving host for cross-pool SXM
We do this in the VGPU import handler when the `live` flag is set. The VM
import handler is called on the receiving host of a cross-pool migration, by
both VM.assert_can_migrate (as a dry-run) and VM.migrate_send (for real).

Note that `vgpu_record` represents the VGPU record that we will create on the
receiving pool, if all goes well. The reference fields in there have already
been resolved for the receiving pool by the import handlers of the relevant
objects.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>