Add Host.allocate_resources_for_vm internal call
This will be used by `VM.migrate_send` to another pool, in order to reserve
memory and GPU resources on the host in the destination pool. It just calls
`VM.allocate_vm_to_host` in `Message_forwarding`, which sets the relevant
`scheduled_to_resident_on` fields. This is the same function that is used by
the other (intra pool) VM lifecycle operations.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>