xenops_interface: factor out PCI address in Vgpu.t
The PCI address (of the pGPU that will be used) is common to all
implementations. We need it to be an explicit field at the level of `Vgpu.t`,
so that we can map an instance of `Vgpu.t` to another PGPU during a
live-migration.

The per-type PCI address fields will no longer be used, but to not break the
upgrade case, the record field is kept, but turned into an option type. The
`t_of_rpc` function is overridden to include upgrade logic to transfer values
to the new field.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>