Check if VM has VDIs with CBT enabled
in Xapi_vm_lifecycle.check_operation_error for the VM.migrate_send
operation.

It is not necessary to duplicate this check for
Xapi_vm_migrate.migrate_send, as this newly-added check is performed for
migrate_send in the message forwarding layer by calling
with_vm_operation.

We have to protect the Db lookup involving the VDI with a try/catch
block, because VBDs representing empty CDs have a null VDI field, and
that would result in an error in this function, and during the
VM.update_allowed_operations function, which calls this, or when we
check the validity of VM operations in the message forwarding layer,
where this function is used indirectly.

In general, all VDI, VBD, and other database accesses should be
protected with try/with in this function, as this function uses a
snapshot of the VM record, so the references in the record might become
invalid by the time we use them.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>