Rethrow VDI_INCOMPATIBLE_TYPE error during VM.revert
instead of ignoring the error and returning the more general
vm_revert_failed error code. This will ensure that when some of the VDIs
of the snapshot are cbt_metadata, the user will see this more specific
error.

This exception can come from a lower-level VDI operation called during
VM.revert. Therefore it is not necessary to explicitly check for
cbt_metadata VDIs during VM.revert, as the VDI operations that we run
should do this check, and in case of an error,
Xapi_vm_clone.safe_clone_disks should clean up after the interrupted
operation.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>