Add VDI.set_on_boot to vdi_operations enum
To prevent a possible race between set_on_boot (to reset) and
enable_cbt, adding set_on_boot as an operation ensures exclusivity. We
want to prevent having VDI.on_boot = reset and CBT enabled at the same
time, because in that case the CBT metadata will become incorrect.

The set_on_boot operation is always disallowed now if CBT is enabled,
regardless of its parameter (thus even setting on_boot to persist is
disallowed), but that is fine, because with the current
checks, we cannot get into a state when on_boot is set to reset and CBT
is enabled.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>