CA-255133: Pool.eject - Don't write bridge into inventory
On Pool.eject, xcp-networkd startup creates bridges if management
bridge is present in inventory.

Then firstboot again tries to create PIFs and bridges.
This is a failure for management vlan case.

To fix this, we can avoid writing management bridge into inventory
and let the firstboot create the management pif and update the inventory.

Signed-off-by: Sharad Yadav <sharad.yadav@citrix.com>