HFX-2137 CA-254515: Check database when updating VDI.snapshot_of
The VDI snapshot's `snapshot_of` field would become a null reference if
its name_label or name_description field was changed, or if vdi-update
was called on it. This was happening because it was only checking the
VDI map argument passed into the `find_vdi` function, which did not
contain the VDI that was originally snapshoted.

Now, if the map argument does not contain the snapshoted VDI, `find_vdi`
will check the database for the snapshoted VDI, which it should find if
it still exists, and hence will no longer return a null reference and
will retain its original `snapshot_of` value.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>