HFX-2137 CA-254515: Update new VDIs after they are created
Potential problem:
This should resolve the `snapshot_of` field becoming NULL in a
situation where the snapshoted VDI and the snapshot VDI are both
newly-created and their order of creation can be anything (which
would have been a problem if the snapshot VDI is created before the
snapshoted VDI, because then when setting the snapshot VDI's
`snapshot_of` field it would not be able to find the snapshoted VDI
in the database because it has not yet been created, and would hence
set the snapshot's `snapshot_of` field to null).

Solution:
Move the `to_update` definition to after `db_vdi_map` is updated with
the newly-created VDIs; that way we end up with the new VDIs inside
the `to_update` map with all the information we need, including the
VDI's reference (from DB), record (from DB) and info (from SR scan).
The newly-created VDIs are then iterated through to update all its
fields, including their `snapshot_of` field which may have been
wrongfully set to null.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>