CP-22597: datamodel: update `migrate_send` and `assert_can_migrate` calls
Introduces a new `vgpu_map` parameter mapping source vgpus to destination
vgpu_groups.  The parameter comes with an empty map as default.

Note that due to its recent addition, we are forced to add it as last
parameter. Adding it next to the other maps will trigger a compilation
error.

This commit introduces a fake release that will have to be replaced with
the correct one once known.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>