VDI.snapshot: inherit cbt_enabled field from snapshotted VDI
Currently this field is not part of vdi_info, which is the information
about a VDI that a SMAPIv2 storage backend can provide. Therefore we
have to set this explicitly during the XenAPI call, as
Xapi_sr.update_vdis will not update this field. This database update
should be removed once we add the cbt_enabled property to vdi_info, then
this field will be updated automatically by Xapi_sr.update_vdis.

Since the cbt_enabled property isn't in vdi_info, this also means that
in Xapi_sr.update_vdis, for VDIs that are returned by the storage
backend but aren't yet in the database, we will create them in the Db
with the cbt_enabled field set to false, as we do not know its actual
value.

This commit does not change the cbt_enabled field for copied VDIs, the
cbt_enabled field of the copy should be false.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>