Implement VDI.data_destroy XenAPI call
This call is idempotent, therefore it is allowed on a VDI of type
cbt_metadata.

Repeat the same checks done for VDI.destroy in check_operation_error, in
addition to the checks specific to data_destroy. However, some of the
checks performed by VDI.destroy are irrelevant to VDI.data_destroy,
because CBT can only be enabled on VDIs of type user or system,
therefore we can never get into a state where CBT is enabled for a
ha_statefile, redo_log, or metadata VDI.
But I still decided to extract the VDI.destroy checks and call these for
both operations, to avoid the various checks getting out of sync - it is
safe to still perform these irrelevant checks for VDI.data_destroy.

I have also reused the implementation of VDI.destroy, and this means
that VDI.data_destroy suffers from the same race as VDI.destroy: we
delete the VBDs referring to the VDI shortly after (data_)destroying the
VDI.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>