VDI.set_metadata_of_pool: only allow `metadata VDI
When we run this function from Xapi_sr.find_or_create_metadata_vdis, we
first set the type of the VDI passed to it to `metadata, so this check
is safe to add. Also, in the various VDI handling functions in Xapi_sr
and Xapi_dr, we only consider VDIs of type `metadata, so we should
enforce that a VDI that contains the metadata of a pool has type
`metadata.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>