CA-242028: Remove exauth pbis cache mechanism.
This is a cache-related issue. Now we have two kinds of cache mechanism in PBIS:
- xapi code in extauth_plugin_ADpbis.ml named `exauth cache`
- PBIS native cache mechanism /opt/pbis/bin/config --detail CacheEntryExpiry

It is not necessary to have two cache mechanism at the same time.

It is safe to remove the `exauth PBIS cache`. Removing the cache will not cause a regression, seeing test [https://info.citrite.net/display/eng/Deeper+performance+test+for+PBIS].

Signed-off-by: Wei Xie <wei.xie@citrix.com>