CA-226886: VDI import: handle exn if SR is unavailable for any host.
This commit adds proper handling in VDI import for exceptions raised
when the SR is not accessible by any host. For example, this can
happen during an update-upload operation when trying to use the
default SR but it was down.

The exception handler marks the relevant task as failed, and returns
an appropriate http error code to the client.

Factor out duplication of finding task_id and subtask in request.

Signed-off-by:Min Li <min.li1@citrix.com>
(cherry picked from commit ba956c6a9142f93064416ae4e676da7ccd958eb5)