Add unit tests for CA-254515
There are 2 scenarios we test; the first is when we update a VDI
snapshot which already exists in the database. The second is when
we are creating a snapshot and its original VDI, in that order
(because they exist in the SR but not the database). In both cases
we test to see that the snapshot's `snapshot_of` field has not
become null.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>