Implement VDI.export_changed_blocks XenAPI call
We require both SRs to be attached for the corner case where we ask for
the changed blocks between two VDIs from two different SRs. However,
this is not strictly necessary: in the above case we just return all the
blocks of the second VDI, so we only need the SR of the second VDI to be
attached.

For this operation, even though we could allow the VDI to be attached as
read-only, we require it to be detached, to simplify the implementation
of the storage plugins - now we have one less thing to consider, and
this makes the implementation more flexible.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>