CP-23026 CA-258536 update update records at startup
This commit changes the behaviour of Xapi_pool_update.resync_host which
reads updates from the file system and is called during startup.
Previously, new update records were created for updates that were
unknown. This commit changes this behaviour: for an update found in the
file system, an existing record is selectively updated.

The rationale for this change is that an updated xapi at startup might
interpret an existing update's meta data differently than before. This
provides an opportunity to correct previous wrong or inconvenient
interpretations. This particular instance was triggered by the
introduction of the enforce-homogeneity attribute in updates.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>