CA-259288 provide descriptive reason when raising pool_hosts_not_homogeneous
With this commit, when pool_hosts_not_homogeneous is raised in pool-join
pre-checks, it provides a descriptive error message rather than a UUID.
This is now in line with other call sites of this error. The main
concern here is that this string is also shown in XenCenter to the user.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>