VDI.data_destroy: reset the VDI's content_id
to a special constant. The content_id of the data_destroyed VDI has to
be different from its old content_id and the content_ids of all the
other VDIs with data, to ensure that we won't select it as the base VDI
during storage motion. Since the data of the VDI has been destroyed, it
should be considered different from all other VDIs that have data.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>