.merlin: search source dir for source files first before _build
Sometimes I go to the definition of a symbol from a different file using
merlin, and edit that file. However, with the old merlin file I jumped
to the source file in the _build directory, and my modifications got
lost after a "make clean".

I changed the order of entries in the .merlin file, so that for existing
source files, we will go to the source directory now instead of _build,
but auto-generated files are still found in the _build directory.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>