HFX-2143 CA-258652/SCTX-2565: Get initial host memory from squeezed
This is a backport of 821bde00310b9ef3f04f2cf33b3901b675fc393e

Prior to this fix, xapi was responsible for calculating the
initial host free memory and recording it in a file in dom0.
This caused an issue on hosts with a CVM; xapi would exclude
the CVM memory in its calculation but it would include it in
subsequent reboots; this happened because, on firstboot, the
CVM is started after xapi, but on reboots it starts before.

squeezed always starts before the CVM and xapi, so the
calculation is moved to squeezed to ensure predictable
behaviour.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>