Test CA-253489: xe update-upload should check default SR
This tests the error handling of the "xe update-upload" CLI command,
that it returns an error if no SR was specified on the command line, and
the pool has no default SR or it is invalid.

Without the CLI-side check, this unit test will fail, even if the fix
for the underlying cause of CA-253489 is there in the import handler,
because it cannot communicate with the HTTP handler. So without the
required check, this test will fail due to timeout.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>