CA-259579: watch balloon-active key to infer ballooning complete
Adding a timeout to avoid watching for an indefinite amount of time.
The watcher is called twice, once (ingnoring any timeout issue) when we
have just performed the ballooning. And once right before issuing the
suspend request.

Issue context:

Before migration we balloon the memory down to dynamic_memory_min. On
windows the driver can sometimes persist ballooning when xenopsd believes
it has completed its job (also it can restart ballooning when additional
memory gets freed). The ballooning driver ignores suspend requests and
this can make the migration timeout leaving the vms in a bad state.

This mitigates the issue by monitoring the ballooning key in xenstore
and issuing the suspend request only after it is done. It also adds a
configurable timeout to cancel the migration if the additional ballooning
is taking too long, raising an error that provides a meaningful suggestion
in xapi (Increase the memory dynamic min value to reduce ballooning).

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>